uni_hub/backend/notifications/serializers.py
# serializers.py
from rest_framework import serializers
from .models import Notification

class NotificationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Notification
        fields = ['id', 'user_email', 'message', 'is_read', 'created_at']  # Added 'user_email'