uni_hub/backend/requirements.txt
django>=5.0.0
django-extensions
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
pillow>=10.1.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.1
drf-nested-routers>=0.93.4 